✨ First promising sound sensor implementation (#1)
Sound voltage is now captured from the ESP's own ADC instead 
of using the ADS1115, which proved to not be up to the task.

Peak to peak envelope is calculated over a sample duration of 
1s, this is moved back to a voltage range of 0-3.3V, then an RMS 
is approximated by multiplying with 0.7071, and finally a sound 
pressure value is calculated based on 
https://electronics.stackexchange.com/a/219721

The resulting dBSPL values still seem too high, I have the suspicion 
the microphone on our MAX4466 boards might not have a 
sensitivity of -44dB like the one on the Adafruit boards, but I was 
not able to source that piece of information. Testing with an offset 
of -15dB applied to the visualizations looks ok-ish.

⚠️ There is a change needed on the boards, connect SND signal pin 
to A0 on ESP (simple bridge cable suffices).